requirements.txt
numpy
Flask
flask-cors
pillow
tensorflow
matplotlib